Split gym (#1154)
* split

* black_and_import

* adding_gym

* move_stuff

* fix_black

* fix_typing

* fix_typo

* pylint
